requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP and API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication and Security
bcrypt==4.1.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Machine Learning and NLP
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Message Queue
celery==5.3.4
kombu==5.3.4
redis[hiredis]==5.0.1

# Blockchain and Crypto
web3==6.12.0
cryptography==41.0.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
bandit==1.7.5

# External API Integrations
wikipedia==1.4.0
googlemaps==4.10.0
google-cloud-vision==3.4.5

# Web Scraping
beautifulsoup4==4.12.2
scrapy==2.11.0

# Web UI
jinja2==3.1.2
python-multipart==0.0.6

# Performance and Load Testing
locust==2.17.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jsonschema==4.20.0
ipaddress==1.0.23

# Production WSGI Server
gunicorn==21.2.0

# Monitoring and Observability
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Additional Security
python-jose[cryptography]==3.3.0
slowapi==0.1.9